+++ pyprep/find_noisy_channels.py
@@ -6,7 +6,7 @@ from scipy import signal
 
 from pyprep.ransac import find_bad_by_ransac
 from pyprep.removeTrend import removeTrend
-from pyprep.utils import _mat_iqr, _mat_quantile, filter_design, madmed
+from pyprep.utils import _filter_design, _mad, _mat_iqr, _mat_quantile
 
 
 class NoisyChannels:
@@ -112,7 +112,7 @@ class NoisyChannels:
         if self.sample_rate <= 100:
             return self.EEGData.copy()
 
-        bandpass_filter = filter_design(
+        bandpass_filter = _filter_design(
             N_order=100,
             amp=np.array([1, 1, 0, 0]),
             freq=np.array([0, 90 / self.sample_rate, 100 / self.sample_rate, 1]),
@@ -225,7 +225,7 @@ class NoisyChannels:
 
         # Detect channels with flat or extremely weak signals
         FLAT_THRESHOLD = 1e-15  # corresponds to 10e-10 µV in MATLAB PREP
-        flat_by_mad = madmed(EEGData, axis=1) < FLAT_THRESHOLD
+        flat_by_mad = _mad(EEGData, axis=1) < FLAT_THRESHOLD
         flat_by_stdev = np.std(EEGData, axis=1) < FLAT_THRESHOLD
         flat_channel_mask = flat_by_mad | flat_by_stdev
         flat_channels = self.ch_names_original[flat_channel_mask]
@@ -312,8 +312,8 @@ class NoisyChannels:
         # < 50 Hz amplitude for each channel and get robust z-scores of values
         if self.sample_rate > 100:
             noisiness = np.divide(
-                madmed(self.EEGData - self.EEGFiltered, axis=1),
-                madmed(self.EEGFiltered, axis=1),
+                _mad(self.EEGData - self.EEGFiltered, axis=1),
+                _mad(self.EEGFiltered, axis=1),
             )
             noise_median = np.nanmedian(noisiness)
             noise_sd = np.median(np.abs(noisiness - noise_median)) * MAD_TO_SD
@@ -396,7 +396,7 @@ class NoisyChannels:
             channel_amplitudes[w, usable] = _mat_iqr(eeg_raw, axis=1) * IQR_TO_SD
 
             # Check for any channel dropouts (flat signal) within the window
-            eeg_amplitude = madmed(eeg_filtered, axis=1)
+            eeg_amplitude = _mad(eeg_filtered, axis=1)
             dropout[w, usable] = eeg_amplitude == 0
 
             # Exclude any dropout chans from further calculations (avoids div-by-zero)
@@ -406,7 +406,7 @@ class NoisyChannels:
             eeg_amplitude = eeg_amplitude[eeg_amplitude > 0]
 
             # Get high-frequency noise ratios for the window
-            high_freq_amplitude = madmed(eeg_raw - eeg_filtered, axis=1)
+            high_freq_amplitude = _mad(eeg_raw - eeg_filtered, axis=1)
             noiselevels[w, usable] = high_freq_amplitude / eeg_amplitude
 
             # Get inter-channel correlations for the window
