+++ pyprep/utils.py
@@ -321,7 +321,7 @@ def _correlate_arrays(a, b, matlab_strict=False):
         return np.diag(np.corrcoef(a, b)[:n_chan, n_chan:])
 
 
-def madmed(x, axis=None):
+def _mad(x, axis=None):
     """Calculate median absolute deviations from the median (MAD) for an array.
 
     Parameters
@@ -343,10 +343,11 @@ def madmed(x, axis=None):
     med = np.median(x, axis=axis)
     if axis == 1:
         med = med.reshape(-1, 1)  # Transposes array to allow subtraction below
-    return np.median(np.abs(x - med), axis=axis)
+    mad = np.median(np.abs(x - med), axis=axis)
+    return mad
 
 
-def filter_design(N_order, amp, freq):
+def _filter_design(N_order, amp, freq):
     """Create FIR low-pass filter for EEG data using frequency sampling method.
 
     Parameters
@@ -394,7 +395,7 @@ def filter_design(N_order, amp, freq):
     return kernel
 
 
-def split_list(mylist, chunk_size):
+def _split_list(mylist, chunk_size):
     """Split list in chunks.
 
     Parameters
@@ -416,7 +417,7 @@ def split_list(mylist, chunk_size):
     ]
 
 
-def print_progress(current, end, start=None, stepsize=1, every=0.1):
+def _print_progress(current, end, start=None, stepsize=1, every=0.1):
     """Print the current progress in a loop.
 
     Parameters
@@ -454,70 +455,7 @@ def print_progress(current, end, start=None, stepsize=1, every=0.1):
             print("{0}%...".format(int(pct)), end=" ", flush=True)
 
 
-def make_random_mne_object(
-    ch_names,
-    ch_types,
-    times,
-    sfreq,
-    n_freq_comps=5,
-    freq_range=[10, 60],
-    scale=1e-6,
-    RNG=np.random.RandomState(1337),
-):
-    """Make a random MNE object to use for testing.
-
-    Parameters
-    ----------
-    ch_names : list
-        names of channels
-    ch_types : list
-        types of channels
-    times : np.ndarray, shape (1,)
-        Time vector to use.
-    sfreq : float
-        Sampling frequency associated with the time vector.
-    n_freq_comps : int
-        Number of signal components summed to make a signal.
-    freq_range : list, len==2
-        Signals will contain freqs from this range.
-    scale : float
-        Scaling factor applied to the signal in volts. For example 1e-6 to
-        get microvolts.
-    RNG : np.random.RandomState
-        Random state seed.
-
-    Returns
-    -------
-    raw : mne.io.Raw
-        The mne object for performing the tests.
-    n_freq_comps : int
-        Number of random frequency components to introduce.
-    freq_range : tuple
-        The low (`freq_range[0]`) and high (`freq_range[1]`) endpoints of
-        a frequency range from which random draws will be made to
-        introduce frequency components in the test data.
-    """
-    n_chans = len(ch_names)
-    signal_len = times.shape[0]
-    # Make a random signal
-    signal = np.zeros((n_chans, signal_len))
-    low = freq_range[0]
-    high = freq_range[1]
-    for chan in range(n_chans):
-        # Each channel signal is a sum of random freq sine waves
-        for freq_i in range(n_freq_comps):
-            freq = RNG.randint(low, high, signal_len)
-            signal[chan, :] += np.sin(2 * np.pi * times * freq)
-
-    signal *= scale  # scale
-
-    # Make mne object
-    info = mne.create_info(ch_names=ch_names, sfreq=sfreq, ch_types=ch_types)
-    raw = mne.io.RawArray(signal, info)
-    return raw, n_freq_comps, freq_range
-
-
-def verify_free_ram(data, n_samples, n_channels, max_prop=0.95):
+def _verify_free_ram(data, n_samples, n_channels, max_prop=0.95):
     """Check if enough memory is free to run RANSAC with the given parameters.
 
     Parameters
